pyproject.toml
[project]
name = "pyopus-smarthome"
version = "0.1.0"
description = "Async Python client for OPUS SmartHome gateway REST API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Philipp Piwo", email = "github@thepiwo.com" }
]
requires-python = ">=3.12"
dependencies = ["aiohttp>=3.9.0"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
keywords = ["opus", "smarthome", "enocean", "home-automation", "aiohttp"]

[project.urls]
Homepage = "https://github.com/thepiwo/pyopus-smarthome"
Repository = "https://github.com/thepiwo/pyopus-smarthome"
Issues = "https://github.com/thepiwo/pyopus-smarthome/issues"

[project.scripts]
opus-qr = "pyopus_smarthome.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
